chore: split error hierarchy
