+++ app/errors/config-error.ts
@@ -1,0 +1,8 @@
+import { MESSAGES } from '../constants.js'
+import { WebSSH2Error } from './webssh2-error.js'
+
+export class ConfigError extends WebSSH2Error {
+  constructor(message: string) {
+    super(message, MESSAGES.CONFIG_ERROR)
+  }
+}
