+++ app/errors/ssh-connection-error.ts
@@ -1,0 +1,8 @@
+import { MESSAGES } from '../constants.js'
+import { WebSSH2Error } from './webssh2-error.js'
+
+export class SSHConnectionError extends WebSSH2Error {
+  constructor(message: string) {
+    super(message, MESSAGES.SSH_CONNECTION_ERROR)
+  }
+}
